create source-code.C
Three major operations can be done in this calendar application. To find out the day corresponding to a given date, the date, month and year are asked. You can list the days and dates of any month of any year. For example, entering 21 10 2019 (21st October) will give you an output as “MONDAY”. The third feature of this C mini project on Calendar application utilizes file handling. With this feature, you can add important notes with corresponding dates. The functions used in the source code are simple and easy to understand. It provides very simple interface. 
It’s a Menu driven program, so list of options is provided where whatever user chooses to opt for.    
1. FIND THE DAY. 
2. PRINT A CALENDAR. 
3. ADD NOTES. 
4. EXIT